+++ golem-worker-service-base/src/expression/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) use expr::*;
+pub(crate) use text::*;
 mod expr;
-mod reader;
+mod text;
 mod writer;
